Match native frame cadence and W2 fp32 cast in fp64 OUTPUT writer
Two byte-parity gaps in OUTPUT/H2U2V2.OUT and OUTPUT/ZUV.OUT for
F2_24K_fp64 step>=900 traced to writer asymmetries with native.

  1. Native runStepsWithOutput emits an end-of-day frame whenever
     kt+1 == steps_per_day (and (day+1) % NTOUTPUT == 0), in
     addition to the forced final-frame write at s == n. The Taichi
     side previously wrote only the initial frame (jt=0, kt=1) and
     the final frame at the requested checkpoint, so for steps that
     crossed at least one EOD boundary (e.g. 900 = 1 day on F2_24K)
     the Taichi files were short by the EOD frames. The diff for
     step=900 showed ~900 missing all-zero "padding" lines in
     H2U2V2.OUT, which was actually a missing whole frame.

     Fix: dump_taichi_multi_step's on_step callback now writes an
     EOD frame for every (still-open) writer whose checkpoint is
     beyond the current step, in addition to the final frame at
     the writer's own checkpoint. Both checkpoint and EOD use
     state captured in-process via to_numpy() at the matching
     post-sync point. NTOUTPUT is hardcoded to 1 to mirror the
     CLI override path used for alignment runs.

  2. Native writes W2 as ``float`` (mesh.cpp:715), so
     sqrt(U*U + V*V) is fp32-truncated before the 4-decimal
     formatter; Taichi was writing the value at fp64. For W2
     values right at the 4-decimal half-way point the rounding
     direction flipped on a handful of cells. Cast the Python
     value through float32 before formatting so both sides emit
     the same fp32-rounded 4-decimal text.

After the fix, F2_24K_fp64 step=900 OUTPUT files report
text_match=True for H2U2V2.OUT, SIDE.OUT, TIMELOG.OUT;
ZUV.OUT diff_lines drops from 11 to 10 with max_w_diff = 0
(only a residual max_fi_diff = 1e-4 from FI rounding, tracked
separately as an Open Issue).

+++ taichi_impl/output_writer.py
@@ -332,7 +332,10 @@ class OutputWriter:
         u_clamped = np.where(wet, U, 0.0)
         v_clamped = np.where(wet, V, 0.0)
         z_clamped = np.where(wet, Z, self.zbc)
-        w2 = np.where(wet, np.sqrt(U * U + V * V), 0.0)
+        # Native writes W2 as ``float`` (mesh.cpp:715), so the
+        # fp32 truncation may flip 4-decimal rounding for values
+        # right at the half-way point. Mirror the cast here.
+        w2 = np.where(wet, np.sqrt(U * U + V * V).astype(np.float32).astype(np.float64), 0.0)
 
         jt2 = jt if kt == 1 else jt + 1
